# docs/swagger.yaml
basePath: /api
definitions:
  model.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.CreatePostRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      cover_url:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.PostStatus'
        description: draft / published，默认 published
      summary:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        type: string
    required:
    - category_id
    - content
    - title
    type: object
  model.CreateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  model.ListResponse:
    properties:
      data: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  model.Post:
    properties:
      author:
        $ref: '#/definitions/model.User'
      author_id:
        type: integer
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      content:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      like_count:
        type: integer
      status:
        $ref: '#/definitions/model.PostStatus'
      summary:
        type: string
      tags:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  model.PostStatus:
    enum:
    - draft
    - published
    type: string
    x-enum-varnames:
    - PostStatusDraft
    - PostStatusPublished
...
  model.Tag:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.UpdatePostRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      cover_url:
        type: string
      status:
        $ref: '#/definitions/model.PostStatus'
      summary:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        type: string
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserRole:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleUser
  utils.Response:
    properties:
      code:
        $ref: '#/definitions/utils.ResponseCode'
      data: {}
      message:
        type: string
    type: object
  utils.ResponseCode:
    enum:
    - 0
    - 400
    - 401
    - 403
    - 404
    - 500
    - 1001
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeInternalError
    - CodeBusinessError
host: localhost:8080
info:
  contact:
    name: 博客作者
    url: http://localhost:8080
  description: 基于 Gin + GORM + MySQL + Redis 的个人博客后端 RESTful API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:8080
  title: 个人博客 API
  version: "1.0"
paths:
...
  /comments/{id}/like:
    get:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    count:
                      type: integer
                    liked:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取评论点赞状态
      tags:
      - 评论点赞
    post:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    count:
                      type: integer
                    liked:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 切换评论点赞状态
      tags:
      - 评论点赞
  /comments/likes:
    get:
      parameters:
      - description: 评论 ID，逗号分隔，例如 1,2,3
        in: query
        name: ids
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties:
                    properties:
                      count:
                        type: integer
                      liked:
                        type: boolean
                    type: object
                  type: object
              type: object
      summary: 批量获取评论点赞状态
      tags:
      - 评论点赞
  /posts:
    get:
      parameters:
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认 10
        in: query
        name: page_size
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - description: 分类 ID
        in: query
        name: category_id
        type: integer
      - description: 标签 ID
        in: query
        name: tag_id
        type: integer
      - description: 状态：draft/published
        in: query
        name: status
        type: string
      - description: 排序：created_at/view_count
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.ListResponse'
              type: object
      summary: 获取文章列表
      tags:
      - 文章
    post:
      consumes:
      - application/json
      parameters:
      - description: 文章信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建文章
      tags:
      - 文章
  /posts/{id}:
    delete:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除文章
      tags:
      - 文章
    get:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Post'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取文章详情
      tags:
      - 文章
    put:
      consumes:
      - application/json
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文章更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新文章
      tags:
      - 文章
...
  /posts/{id}/like:
    get:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    count:
                      type: integer
                    liked:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取文章点赞状态
      tags:
      - 点赞
    post:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    count:
                      type: integer
                    liked:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 切换文章点赞状态
      tags:
      - 点赞
  /posts/hot:
    get:
      parameters:
      - description: 数量限制，默认 10
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Post'
                  type: array
              type: object
      summary: 获取热门文章
      tags:
      - 文章
  /tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Tag'
                  type: array
              type: object
      summary: 获取标签列表
      tags:
      - 标签
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签名称
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Tag'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建标签（管理员）
      tags:
      - 标签
  /tags/{id}:
    delete:
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除标签（管理员）
      tags:
      - 标签
...
securityDefinitions:
  BearerAuth:
    description: 请输入 "Bearer {token}"，例如：Bearer eyJhbGciOiJIUzI1NiIs...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"